changed data types in uzivatelDTO to fit new models

+++ DopravniPodnik/DopravniPodnik/Data/DTO/UzivatelDTO.cs
@@ -4,8 +4,6 @@ namespace DopravniPodnik.Data.DTO;
 
 public class UzivatelDTO
 {
-    // uzivatel
-    // public decimal id_uzivatele { get; set; }
     public string uzivatelske_jmeno{ get; set; }
     public string heslo{ get; set; }
     public string jmeno{ get; set; }
@@ -13,20 +11,19 @@ public class UzivatelDTO
     public DateTime datum_zalozeni{ get; set; }
     public DateTime datum_narozeni{ get; set; }
     //typ_uzivatele
-    // public decimal id_typ_uzivatele{ get; set; }
     public string nazev_typ_uzivatele{ get; set; }
     //foto
-    public decimal id_foto{ get; set; }
+    public int id_foto{ get; set; }
     public string foto_jmeno_souboru{ get; set; }
     public Image foto_data{ get; set; }
     public DateTime foto_datum_pridani{ get; set; }
     //adresa
-    public decimal id_adresy{ get; set; }
+    public int id_adresy{ get; set; }
     public string mesto{ get; set; }
     public string ulice{ get; set; }
     public short cislo_popisne{ get; set; }
     //possible relations
-    public decimal? id_zakaznika{ get; set; }
-    public decimal? id_zamestnance{ get; set; }
+    public int? id_zakaznika{ get; set; }
+    public int? id_zamestnance{ get; set; }
     
 }
